+++ lib/cable_ready.rb
@@ -15,6 +15,7 @@ require "cable_ready/identifiable"
 require "cable_ready/operation_builder"
 require "cable_ready/config"
 require "cable_ready/broadcaster"
+require "cable_ready/broadcastable"
 require "cable_ready/sanity_checker"
 require "cable_ready/compoundable"
 require "cable_ready/channel"
